# Conflicts:
#	.project
#	cnf/central.maven
#	cnf/local/index.xml
#	cnf/local/index.xml.sha
#	cnf/release/index.xml
#	cnf/release/index.xml.sha
#	io.openems.backend.edgewebsocket.api/.settings/org.eclipse.core.resources.prefs
#	io.openems.backend.edgewebsocket.impl/.settings/org.eclipse.core.resources.prefs
#	io.openems.backend.metadata.api/.settings/org.eclipse.core.resources.prefs
#	io.openems.backend.timedata.api/.settings/org.eclipse.core.resources.prefs
#	io.openems.backend.uiwebsocket.api/.settings/org.eclipse.core.resources.prefs
#	io.openems.common/src/io/openems/common/timedata/CommonTimedataService.java
#	io.openems.edge.application/EdgeApp.bndrun

+++ io.openems.edge.ess.api/src/io/openems/edge/ess/dccharger/api/EssDcCharger.java
@@ -103,7 +103,7 @@ public interface EssDcCharger extends OpenemsComponent {
 	 * 
 	 * @return the Channel
 	 */
-	default Channel<Integer> getActualEnergy() {
+	default Channel<Long> getActualEnergy() {
 		return this.channel(ChannelId.ACTUAL_ENERGY);
 	}
